+++ internal/bot/components/misc/misc.go
@@ -88,6 +88,14 @@ func Init(kit *route.Kit, _ *state.State) error {
 		Category: meta.CategoryMisc,
 	})
 
+	kit.AddCommand(&route.Command{
+		Name:        "Dump your data",
+		Help:        "Get a dump of all the data associated with your Discord user ID in the bot database",
+		CommandText: []string{"mydata"},
+		Run: comp.GetMyData,
+		Category: meta.CategoryMisc,
+	})
+
 	return nil
 
 }