+++ .github/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Added initial selection to bios (**BREAKING**)
   * For accounts with multiple bios registered, the bio to start at in the carousel is now chosen by the user
 * `<p>forgetme` command
+* `<p>mydata` command
 ### Changed
 * Switch to using hashed user IDs in kick, ban and verification fail tables (**BREAKING**)
   * This means this data can be retained while complying with a request from a user to delete all identifying data